Functional but ugly.
Successfully generates and uses classes spun at jlink time for
lambda implementation.  The indy is converted to a call to a "thunk"
method of the same signature that handles allocating the Interface
implementation and repushing the args.

This introduces an extra stackframe in the exception stacktrace.
Likely this can be avoided by more argument shuffling in the caller
but didn't seem worth it given this is the allocation, not call path.

Nest mates are fixed up.
This is a user visible change in both the generated class and in the
class that defines the lambda.  They must now both be part of the
same nest.  User code can observe this in strange places.
ie: class Out { class Inner { void m() { Runnable r=()->{}; } } }
Will result in the Inner$$Lambda$$123 being a member of Out's nest.

Hidden classes
The original LambdaMetaFactory uses hidden classes for the lambda
implementation.  This is a prespun class loaded in the normal way
which means it will not be a hidden class.

Caller-sensitive stack walks?
Need to determine what the effect of adding this addition frame for
lambda generation is on caller-sensitive stackwalks.