Prototype Holding and SingleAttributeTransforms
HoldingTransform finds a ClassElement based on a filter and
doesn't pass it to the builder.  It instead provides a "get()" method
that allows later stages of the pipeline to pull the value out of the
HoldingTransform instance.  Must be used with ofStateful

SingleAttributeTransform is for dealing with Attributes that have the
"Highlander" (there can only be one) property.  It allows users to
provide two lambdas for either updating or creating the Attribute.

The SingleAttribute version is cleaner than the HoldingTransform
but I'm not clear yet on how it will compose.